Refactor
The directory layout was completely flattened so that src doesn't
have any subdirectories now. What used to be called 'provinces'
have now been broken into a Clickable component and a Question
class. The provinceQuiz state was also changed to have the fields
currentQuestion, otherQuestions, and finished, which make it much
more self documenting.

In cleaning all of that up, a few logic errors became apparent, and
fixing those streamlined a few uglier parts of the code base.